Update chapter 4.1

+++ сhapter4/chapter4.1/App.cpp
@@ -5,7 +5,7 @@ enum class MonsterType { ORGE, GOBLIN, SKELETON, ORC, TROLL };
 struct Monster {
   MonsterType type; // Тип
   std::string name; // Ім'я
-  uint16_t health;  // Здоров'я
+  int health;       // Здоров'я
 };
 
 std::string getMonsterType(const MonsterType &monsterType) {
